Refactor utility functions to utils.py
